+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+## Fixed
+- Apollo: Prominence Flare/Glorious Disaster (thanks ParanoidDrone)
+- Hera: Dying Wish should properly display now (thanks ParanoidDrone)
 
 ## [2.8.21] - 2025-11-02